Moved new recombination code to "combine.py"
