+++ tests/test_combine.py
